Added possibility to automatically export variables on provision to env file. This can be used in Vagrant by doing `config.user.auto_export = <format>`
